Create package dir for each applications, saperate own routes, forms, create new configuration dir for replace the app var, added apps name ot the url, reimport each library package yo each application and create own __init__ files, delete the main routes and form
