+++ flask_pro/flaskpro/templates/navbar.html
@@ -10,30 +10,32 @@
 						<span class="icon-bar"></span>
 						<span class="icon-bar"></span>
 					</button>
-	             	<a class="navbar-brand" href="{{ url_for('home') }}">Flask Pro</a>
+	             	<a class="navbar-brand" href="{{ url_for('settings.home') }}">Flask Pro</a>
 	            </div>
 	            <div id="navbar" class="navbar-collapse collapse">
 	              <ul class="nav navbar-nav">
-	                <li class="active"><a href="{{ url_for('home') }}">Home</a></li>
-	                <li><a href="/about">About</a></li>
+	                <li class="active"><a href="{{ url_for('settings.home') }}">Home</a></li>
+	                <li><a href="{{ url_for('settings.about') }}">About</a></li>
 	                <li><a href="/contact">Contact</a></li>
 
 	               
-		            <li><a href="{{ url_for('post_list') }}">Posts</a></li>
+		            <li><a href="{{ url_for('posts.post_list') }}">Posts</a></li>
 		            
 
 	                <li class="dropdown">
 	                  <a href="" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">User <span class="caret"></span></a>
 	                  <ul class="dropdown-menu">
 	                  	{% if current_user.is_authenticated %}
-	                  		<li><a href="{{ url_for('logout') }}">Logout</a></li>
+	                  		<li><a href="{{ url_for('accounts.logout') }}">Logout</a></li>
 	                    	<li role="separator" class="divider"></li>
 	                    	<li class="dropdown-header">Setting</li>
-	                    	<li><a class="nav-item nav-link" href="{{ url_for('account') }}">Account</a></li>
-	                    	<li><a class="nav-item nav-link" href="/change_password">Change Password</a></li>
+	                    	<li><a class="nav-item nav-link" href="{{ url_for('accounts.account') }}">Account</a></li>
+	                    	<li><a class="nav-item nav-link" href="/change_password">
+	                    		Change Password
+	                    	</a></li>
 	                    {% else %}
-	                    	<li><a href="{{ url_for('user_login') }}">Login</a></li>
-	                    	<li><a href="{{ url_for('user_register') }}">Register</a></li>
+	                    	<li><a href="{{ url_for('accounts.user_login') }}">Login</a></li>
+	                    	<li><a href="{{ url_for('accounts.user_register') }}">Register</a></li>
 	                    {% endif %}
 	                  </ul>
 	                </li>
