+++ flask_pro/flaskpro/templates/posts.html
@@ -9,8 +9,8 @@
 		</h3>
 		{% for post in posts.items %}
 			<img src="{{ url_for('static', filename='images/profile_picture/' + post.author.user_image) }}" class="rounded img-thumbnail float-right" >
-			<h4><a href="{{ url_for('post_detail', post_id=post.id) }}">{{ post.title }}</a></h4>
-			<a href="{{ url_for('user_posts', username=post.author.username) }}">{{post.author.username}}</a>
+			<h4><a href="{{ url_for('posts.post_detail', post_id=post.id) }}">{{ post.title }}</a></h4>
+			<a href="{{ url_for('posts.user_posts', username=post.author.username) }}">{{post.author.username}}</a>
 			<strong>{{ post.created_date.strftime('%h, %d %Y | %H:%m') }}</strong>
 			<p class="text-justify">{{ post.content }}</p>
 		{% endfor %}
@@ -18,11 +18,11 @@
 		{% for page_num in posts.iter_pages(left_edge=1, right_edge=1, left_current=1, right_current=2) %}
 			{% if page_num %}
 				{% if posts.page == page_num %}
-					<a href="{{ url_for('post_list', page=page_num) }}" class="btn btn-primary mb-4">
+					<a href="{{ url_for('posts.post_list', page=page_num) }}" class="btn btn-primary mb-4">
 						{{ page_num }}
 					</a>
 				{% else %}
-					<a href="{{ url_for('post_list', page=page_num) }}" class="btn btn-outline-primary mb-4">
+					<a href="{{ url_for('posts.post_list', page=page_num) }}" class="btn btn-outline-primary mb-4">
 						{{ page_num }}
 					</a>
 				{% endif %}
