+++ flask_pro/flaskpro/templates/registration/login_form.html
@@ -50,13 +50,13 @@
        <br>
         <div class="col-md-4 col-md-offset-4">
           <p class="text-muted">
-            <a href="{{ url_for('reset_request') }}" title="Forget password">Forget Password</a>
+            <a href="{{ url_for('accounts.reset_request') }}" title="Forget password">Forget Password</a>
           </p>
         </div>
     </form>
     <div class="col-md-4 col-md-offset-4">
         <p class="text-muted">
-            To create new account go to register form <a href="{{ url_for('user_register') }}" title="">Register</a>
+            To create new account go to register form <a href="{{ url_for('accounts.user_register') }}" title="">Register</a>
         </p>
     </div>
 </div>
