+++ flask_pro/run.py
@@ -1,5 +1,7 @@
-from flaskpro import app
+from flaskpro import create_app
 
+
+app = create_app()
 
 if __name__ == '__main__':
     app.run(debug=True)
